+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2023-11-09
+### Added
+- GeoPoint
+  - Find Closest Intersection
+
+### Fixed
+- Perl Dependency Fetching
+- GeoPoint
+  - Distance Haversine formula not working for points on opposite ends of globe.
+
 ## [1.1.2] - 2023-10-15
 ### Fixed
 - Arc Course Calculation
@@ -201,7 +211,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Math Tools
   - Conversion Functions
   - Legendre Spherical Harmonics Functions (For Magnetic Calculations)
-
+  
+[1.2.0]: https://github.com/997R8V10/aviation-calc-util/compare/v1.1.2...v1.2.0
 [1.1.2]: https://github.com/997R8V10/aviation-calc-util/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/997R8V10/aviation-calc-util/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/997R8V10/aviation-calc-util/compare/v1.0.20...v1.1.0